backend/requirements.txt
fastapi[standard]<1.0.0,>=0.114.2
httpx<1.0.0,>=0.25.1
pydantic-settings<3.0.0,>=2.2.1
pytest-asyncio<0.24.0,>=0.23.0
pytest<8.0.0,>=7.4.3
pytest-mock<4.0.0,>=3.12.0
